Feat: Add unset/replace option for configs merging
This patch provides the support of '*_unset' and '*_replace' keys
depending the expected behavior. Those keyword must be appended at the
end of the key, and it works only for hashes (and possibly arrays
later).

If the replace keyword is used, then inheritance is disabled.
If the unset keywork is used, then this value will be deleted. Be
careful to not delete core objects, as it could break vagrant.