Change: Update flavors to fit to AWS semantic
This is to help user to use stanard keywords instead of
reinvent the wheel. I use AWS as standard, because it was the
first on the market, and probably the most known. However the
flavor settings have smaller values as these VMs are meant to
be run on developper host, which may not have 48 cpus and
128Gb RAM.