Change(.gitignore): Remove conf/vagrant.yml from tracked files
This changes was made to help the beta users which are not so
comfortable with git to be able to do a git pull without having
conflict with their own conf/vagrant.yml. To track your configuration
file , just do a `git add -f conf/vagrant.yml` to get back the old
behavior. This change will be removed when easy-vagrant will be released
as stable.