Fix(playbook.yml ): network issue with RHEL based boxes
This commit fix a bug when loosing network connexion with RHEL based
boxes after a shortwhile time. This is not am Easy-vagrant bug, but more
a Vagrant or box issue, as the configuration file for the default
network interface is not created when the VM is instanciated.

The fix has been implemented with Ansible, but after thinking a bit, it
should have been done via a basic shell script provisionning. So this could
change later...